Move the data structures to data_structures.py.
