+++ wry/common.py
@@ -204,78 +204,3 @@ def invoke_method(service_name, method_name, options, client, resource_name=None
         raise exceptions.NonZeroReturn(return_value)
     return not return_value
 
-
-class RadioButtons(object):
-    def __init__(self, *values):
-        self.values = values
-        self._selected_value = Ellipsis
-
-    def __repr__(self):
-        out = []
-        for value in self.values:
-            if value == self._selected_value:
-                out.append('<%r>' % value)
-            else:
-                out.append(value.__repr__())
-        return ' | '.join(out)
-
-    def __str__(self):
-        return self.__repr__()
-
-    @property
-    def selected(self):
-        return self._selected_value
-
-    @selected.setter
-    def selected(self, value):
-        if value in self.values:
-            self._selected_value = value
-        else:
-            raise TypeError('%r is an invalid value. Choose one of %r.' % (value, self.values))
- 
-class ToggleButtons(object):
-    def __init__(self, *values):
-        ''' Might want to stop people providing None.'''
-        self.values = values
-        self._selected_values = [Ellipsis, ]
-
-    def __repr__(self):
-        out = []
-        for value in self.values:
-            if value in self._selected_values:
-                out.append('<%r>' % value)
-            else:
-                out.append(value.__repr__())
-        return ' | '.join(out)
-
-    def __str__(self):
-        return self.__repr__()
-
-    def __iadd__(self, value):
-        if value not in self._selected_values:
-            self.toggle(value)
-
-    def __isub__(self, value):
-        if value in self._selected_values:
-            self.toggle(value)
-
-    def toggle(self, value):
-        if value not in self.values:
-            raise TypeError('%r is an invalid value. Choose one of %r.' % (value, self.values))
-        try:
-            self._selected_values.remove(value)
-        except ValueError:
-            self._selected_values.append(value)
-
-    @property
-    def selected(self):
-        return self._selected_values
-
-    @selected.setter
-    def selected(self, values):
-        to_set = []
-        for value in values:
-            if value not in self.values:
-                raise TypeError('%r is an invalid value. Choose one of %r.' % (value, self.values))
-            to_set.append(value)
-        self._selected_values = to_set
